test: merge post and get request, use .sequential
